refactor: Rename Em to Emphasis for consistency.
Also ensure all components are being exported properly, because I 
definitely forget this step.
