+++ src/components/index.js
@@ -1,11 +1,13 @@
 export { default as Nautilus } from './hoc/Nautilus';
 
-export { default as Em } from './inline/Em';
+export { default as Emphasis } from './inline/Emphasis';
 export { default as Strong } from './inline/Strong';
 
 export { default as Button } from './ui/Button';
 export { default as Heading } from './ui/Heading';
+export { default as Icon } from './ui/Icon';
 export { default as Link } from './ui/Link';
 export { default as List } from './ui/List';
 export { default as PageTitle } from './ui/PageTitle';
 export { default as Paragraph } from './ui/Paragraph';
+export { default as TextField } from './ui/TextField';
